--- config.json
@@ -1,6 +1,0 @@
-{
-  "token": "",
-  "prefix": "Strad ",
-  "logsChannel": "419506197847343132",
-  "modRole": "Mentor"
-}
