+++ src/commands/ban.js
