+++ src/commands/daily.js
