+++ src/commands/emit.js
