+++ src/commands/key.js
