+++ src/commands/kick.js
