+++ src/commands/mute.js
