+++ src/functions/isFeedbackable.js
