+++ src/static_data/assets.json
